Add education and experience profile endpoints and refactor error/success handling
Introduced education and experience endpoints to profile, enabling adding, updating, and fetching actions. A refactor was performed to improve error and success handling. Contact information feature was enhanced by offering the opportunities to manage users' contact information and certifications.

+++ server/src/response/success_handling.rs
@@ -4,7 +4,7 @@ use serde::Serialize;
 
 #[derive(Serialize)]
 pub enum AppSuccess {
-    CREATED { id: Option<i32> },
+    CREATED { id: Option<i64> },
     DELETED,
     UPDATED,
 }
